doc: update tutorials

+++ tutorials/ckn_mnist_multi_classification_l1l2.py
@@ -5,9 +5,9 @@ import numpy as np
 
 #load ckn_mnist dataset 10 classes, n=60000, p=2304
 data=np.load('dataset/ckn_mnist.npz')
-y=data['y'].astype("float64")
+y=data['y']
 y = np.squeeze(y)
-X=data['X'].astype("float64")
+X=data['X']
 
 #center and normalize the rows of X in-place, without performing any copy
 preprocess(X,centering=True,normalize=True,columns=False)
